generator/requirements.txt
python-decouple
pymysql
sqlalchemy